Split expressions tests at the module seam (SPEC_0021)
Pure move: lines 1054-end of tests/expressions.rs relocate verbatim into
the new child module tests/expressions/function_calls.rs, keeping the
shared helpers parent-owned. Machine-verified zero non-blank lines lost;
no test removed or weakened. Both files under the SPEC_0021 hard limit.
Focused verification: cargo test -p rumoca-phase-solve (97 passed, 0
failed) on the containing tree; the staged diff is the move alone.
